move SiteConfiguration model to core app + update CoreContextProcessor with site + use site configuration settings
