+++ credentials/apps/core/tests/test_context_processors.py
@@ -1,8 +1,8 @@
 """ Context processor tests. """
 
-from django.contrib.sites.models import Site
 from django.test import override_settings, RequestFactory, TestCase
 
+from credentials.apps.api.tests import factories
 from credentials.apps.core.context_processors import core
 
 
@@ -15,9 +15,13 @@ class CoreContextProcessorTests(TestCase):
 
     @override_settings(PLATFORM_NAME=PLATFORM_NAME)
     def test_core(self):
+        site = factories.SiteFactory()
         request = RequestFactory().get('/')
         request.LANGUAGE_CODE = LANGUAGE_CODE
-        self.assertDictEqual(
-            core(request),
-            {'site': Site.objects.get_current(), 'platform_name': PLATFORM_NAME, 'language_code': LANGUAGE_CODE}
-        )
+        request.site = site
+        expected_output = {
+            'site': site,
+            'platform_name': PLATFORM_NAME,
+            'language_code': LANGUAGE_CODE,
+        }
+        self.assertDictEqual(core(request), expected_output)
