+++ credentials/templates/credentials/base.html
@@ -1,5 +1,6 @@
 {# Base template for credentials-specific pages. #}
 {% extends 'base.html' %}
+{% load compress %}
 {% load i18n %}
 {% load staticfiles %}
 {% block title %}
